FIX(tests): CrÃation de la taxonomie des fallacies informelles
RÃpare les tests dans 'test_informal_definitions.py' qui Ãchouaient Ã cause de la mise Ã jour de Pandas 2.x.

Les problÃ¨mes Ãtaient liÃs aux types de donnÃes utilisÃs pour les clÃs primaires et ÃtrangÃ¨res dans la fixture du DataFrame de taxonomie.

Changements:
- Le type de la clÃ primaire 'PK' a ÃtÃ confirmÃ en 'int64'.
- Le type de la clÃ parente 'parent_pk' a ÃtÃ changÃ en 'float' pour gÃrer correctement les valeurs NaN sans utiliser le type nullable 'Int64' qui causait des 'TypeError' lors des comparaisons.

+++ tests/unit/argumentation_analysis/agents/core/informal/test_informal_definitions.py
@@ -39,8 +39,8 @@ def sample_taxonomy_data():
 def mock_taxonomy_df(sample_taxonomy_data):
     """Fixture pour un DataFrame de taxonomie mocké."""
     df = pd.DataFrame(sample_taxonomy_data)
-    df['PK'] = df['PK'].astype('Int64')
-    df['parent_pk'] = pd.to_numeric(df['parent_pk'], errors='coerce').astype('Int64')
+    df['PK'] = df['PK'].astype('int64')
+    df['parent_pk'] = pd.to_numeric(df['parent_pk'], errors='coerce').astype(float)
     df.set_index('PK', inplace=True)
     return df
